+++ Layout.h
@@ -41,7 +41,7 @@
 
 - (MVNode *)            createDataNode:(MVNode *)parent
                                caption:(NSString *)caption
-                              location:(uint32_t)location
-                                length:(uint32_t)length;
+                              location:(uint64_t)location
+                                length:(uint64_t)length;
 
 @end
